+++ plugins/kiali/src/helper/index.ts
@@ -1,3 +1,4 @@
 export * from './ErrorRate';
+export * from './handlerError';
 export * from './Health';
 export * from './utils';